Fix installation of kas for multi user system
kas needs to be installed with PIPX_HOME pointing to a shared directory,
so other users than root can access it. pipx 1.5.0 allows doing this via
the --global flag, which is unavailable in previous versions and needs a
workaround there. Still, the resulting installation is the same and
files are placed in /opt/pipx, available for all users.

Signed-off-by: Martin Schwan <m.schwan@phytec.de>